fix: weather columns are float

+++ denormalized_trips.sql
@@ -1,7 +1,6 @@
 CREATE TABLE denormalized_trips (
     trip_id                 UInt32,
     vendor_id               String,
-
     pickup_datetime         DateTime,
     dropoff_datetime        Nullable(DateTime),
 
@@ -24,10 +23,10 @@ CREATE TABLE denormalized_trips (
     payment_type            Nullable(String),
     trip_type               Nullable(UInt8),
 
-    precipitation           Nullable(Int8),
-    snow_depth              Nullable(Int8),
-    snowfall                Nullable(Int8),
+    precipitation           Nullable(Float32),
+    snow_depth              Nullable(Float32),
+    snowfall                Nullable(Float32),
     max_temperature         Nullable(Int8),
     min_temperature         Nullable(Int8),
-    average_wind_speed      Nullable(Int8)
+    average_wind_speed      Nullable(Float32)
 ) ENGINE = Log;
